backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
neo4j==5.27.0
pandas==2.2.3
networkx==3.4.2
node2vec==0.5.0
pinecone-client==5.0.1
scikit-learn==1.6.1
httpx==0.28.1
unidecode==1.3.8
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio>=0.24.0
python-louvain==0.16
numpy>=1.26.0
scipy>=1.12.0
networkit>=11.0
ruff==0.9.4
